Fix gzip support
* gzip needs to operate on the mmap object, not
the other way around

* Add gzipped test file

* Add test for gzipped file
