+++ eventio/tests/test_iact.py
@@ -1,19 +1,8 @@
 import eventio
 import pkg_resources
 from os import path
-import numpy as np
 
 from pytest import approx
-
-
-def test_file_has_correct_types():
-    testfile = pkg_resources.resource_filename(
-        'eventio', path.join('resources', 'one_shower.dat')
-    )
-    f = eventio.EventIOFile(testfile)
-    types = [o.header.type for o in f]
-
-    assert types == [1200, 1212, 1201, 1202, 1203, 1204, 1209, 1210]
 
 
 def test_file_has_run_header():
