+++ eventio/tests/test_open_file.py
@@ -32,3 +32,23 @@ def test_file_is_iterable():
     f = eventio.EventIOFile(testfile)
     for event in f:
         pass
+
+
+def test_file_has_correct_types():
+    testfile = pkg_resources.resource_filename(
+        'eventio', path.join('resources', 'one_shower.dat')
+    )
+    f = eventio.EventIOFile(testfile)
+    types = [o.header.type for o in f]
+
+    assert types == [1200, 1212, 1201, 1202, 1203, 1204, 1209, 1210]
+
+
+def test_types_gzipped():
+    testfile = pkg_resources.resource_filename(
+        'eventio', path.join('resources', 'one_shower.dat.gz')
+    )
+    f = eventio.EventIOFile(testfile)
+    types = [o.header.type for o in f]
+
+    assert types == [1200, 1212, 1201, 1202, 1203, 1204, 1209, 1210]
